mods/sendfile/requirements.txt
django-sendfile2~=0.7.2
